auto_mail_sender/requirements.txt
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
httplib2==0.13.0
idna==2.8
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.17.2
oauthlib==3.0.2
pandas==0.25.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.25.3
